chore: rename cmd to commands
